ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
pymongo==4.6.0
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
tensorflow==2.14.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
openai==1.3.5
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
websockets==12.0
celery==5.3.4
flower==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 